refactor: domain 모듈 분리
