+++ app/src/androidTest/java/woowacourse/movie/SeatSelectionActivityTest.kt
@@ -12,9 +12,9 @@ import androidx.test.espresso.matcher.ViewMatchers.isSelected
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.rules.ActivityScenarioRule
+import com.example.domain.Movie
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.movie.domain.Movie
 import woowacourse.movie.view.seatselection.SeatSelectionActivity
 import woowacourse.movie.view.viewmodel.toUIModel
 import java.time.LocalDate
