--- app/src/main/java/woowacourse/movie/domain/DayOfWeek.kt
@@ -1,15 +1,0 @@
-package woowacourse.movie.domain
-
-import java.time.LocalDate
-
-enum class DayOfWeek {
-    WEEKDAY, WEEKEND;
-
-    companion object {
-        private val daysNumber = (1..5)
-        fun checkDayOfWeek(date: LocalDate): DayOfWeek {
-            if (date.dayOfWeek.value in daysNumber) return WEEKDAY
-            return WEEKEND
-        }
-    }
-}
