--- app/src/main/java/woowacourse/movie/domain/Movie.kt
@@ -1,12 +1,0 @@
-package woowacourse.movie.domain
-
-import java.time.LocalDate
-
-data class Movie(
-    val title: String,
-    val startDate: LocalDate,
-    val endDate: LocalDate,
-    val runningTime: Int,
-    val description: String,
-    val moviePoster: Int,
-)
