--- app/src/main/java/woowacourse/movie/domain/policy/DiscountPolicy.kt
@@ -1,9 +1,0 @@
-package woowacourse.movie.domain.policy
-
-import java.time.LocalDateTime
-
-interface DiscountPolicy {
-
-    fun discountPrice(price: Int): Int
-    fun checkPolicy(date: LocalDateTime): Boolean
-}
