--- app/src/main/java/woowacourse/movie/domain/seat/SeatRank.kt
@@ -1,7 +1,0 @@
-package woowacourse.movie.domain.seat
-
-enum class SeatRank(val price: Int) {
-    B(10000),
-    S(15000),
-    A(12000)
-}
