--- app/src/main/java/woowacourse/movie/domain/seat/SeatRow.kt
@@ -1,9 +1,0 @@
-package woowacourse.movie.domain.seat
-
-enum class SeatRow {
-    A, B, C, D, E;
-
-    companion object {
-        fun of(row: Int) = values().getOrNull(row) ?: throw IllegalStateException()
-    }
-}
