+++ domain/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    id("java-library")
+    id("org.jetbrains.kotlin.jvm")
+}
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_7
+    targetCompatibility = JavaVersion.VERSION_1_7
+}
+dependencies {
+    testImplementation("junit:junit:4.13.2")
+}
