+++ domain/src/main/java/com/example/domain/DayOfWeek.kt
@@ -1,0 +1,15 @@
+package com.example.domain
+
+import java.time.LocalDate
+
+enum class DayOfWeek {
+    WEEKDAY, WEEKEND;
+
+    companion object {
+        private val daysNumber = (1..5)
+        fun checkDayOfWeek(date: LocalDate): DayOfWeek {
+            if (date.dayOfWeek.value in daysNumber) return WEEKDAY
+            return WEEKEND
+        }
+    }
+}
