+++ domain/src/main/java/com/example/domain/Movie.kt
@@ -1,0 +1,12 @@
+package com.example.domain
+
+import java.time.LocalDate
+
+data class Movie(
+    val title: String,
+    val startDate: LocalDate,
+    val endDate: LocalDate,
+    val runningTime: Int,
+    val description: String,
+    val moviePoster: Int,
+)
