+++ domain/src/main/java/com/example/domain/Ticket.kt
@@ -1,0 +1,11 @@
+package com.example.domain
+
+import com.example.domain.seat.SeatPosition
+import java.time.LocalDateTime
+
+data class Ticket(
+    val price: Int,
+    val date: LocalDateTime,
+    val numberOfPeople: Int,
+    val seats: List<SeatPosition>
+)
