+++ domain/src/main/java/com/example/domain/policy/DiscountPolicy.kt
@@ -1,0 +1,9 @@
+package com.example.domain.policy
+
+import java.time.LocalDateTime
+
+interface DiscountPolicy {
+
+    fun discountPrice(price: Int): Int
+    fun checkPolicy(date: LocalDateTime): Boolean
+}
