+++ domain/src/main/java/com/example/domain/seat/SeatRank.kt
@@ -1,0 +1,7 @@
+package com.example.domain.seat
+
+enum class SeatRank(val price: Int) {
+    B(10000),
+    S(15000),
+    A(12000)
+}
