+++ domain/src/main/java/com/example/domain/seat/SeatRow.kt
@@ -1,0 +1,9 @@
+package com.example.domain.seat
+
+enum class SeatRow {
+    A, B, C, D, E;
+
+    companion object {
+        fun of(row: Int) = values().getOrNull(row) ?: throw IllegalStateException()
+    }
+}
